topic/#19

+++ src/main/java/com/example/management/model/Customer.java
@@ -32,5 +32,5 @@ public class Customer {
 	
 	private LocalDateTime updatedAt; // 更新日時
 	
-	private Integer invalid; // 削除フラグ
+	private Short invalid; // 削除フラグ
 }
